gcp/functions/cleanup/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.11.*
google-cloud-storage==2.10.*
google-cloud-logging==3.5.*
flask==3.0.*
